requirements.txt
streamlit
python-dotenv
openai
azure-cognitiveservices-vision-computervision
msrest
Pillow